--- a/ext/http/00_serve.ts
+++ b/ext/http/00_serve.ts
@@ -139,7 +139,6 @@ class InnerRequest {
   #upgraded;
   #urlValue;
   #completed;
-  #abortController;
   request;
 
   constructor(external, context) {