--- a/ext/web/03_abort_signal.js
+++ b/ext/web/03_abort_signal.js
@@ -8,7 +8,6 @@ const {
   ArrayPrototypeEvery,
   ArrayPrototypePush,
   ObjectPrototypeIsPrototypeOf,
-  SafeArrayIterator,
   SafeSet,
   SafeSetIterator,
   SafeWeakRef,